Separate items count select to own component
